pulling test_missions out.
